Support more than one record
Turns out I had missed what record count meant in the spec, and stopped reading data after the first record. This adds support for reading all records. It also adds Edf#duration as a way to find the total duration of the data (in millisecond to match how timestamps are stored in JS).